feat(client):templateSelectField
